Bring the command up to date
