+++ .distignore
@@ -1,0 +1,18 @@
+.DS_Store
+.git
+.gitignore
+.gitlab-ci.yml
+.editorconfig
+.travis.yml
+behat.yml
+circle.yml
+phpcs.xml.dist
+phpunit.xml.dist
+bin/
+features/
+utils/
+*.zip
+*.tar.gz
+*.swp
+*.txt
+*.log
