+++ .github/ISSUE_TEMPLATE
@@ -1,0 +1,11 @@
+<!--
+
+Thanks for creating a new issue!
+
+Found a bug or want to suggest an enhancement? Before completing your issue, please review our best practices: https://make.wordpress.org/cli/handbook/bug-reports/
+
+Need help with something? GitHub issues aren't for general support questions, but there are other venues you can try: https://wp-cli.org/#support
+
+You can safely delete this comment.
+
+-->
