+++ .gitignore
@@ -1,3 +1,13 @@
 .DS_Store
+wp-cli.local.yml
 node_modules/
 vendor/
+*.zip
+*.tar.gz
+*.swp
+*.txt
+*.log
+composer.lock
+phpunit.xml
+phpcs.xml
+.phpcs.xml
