--- .travis.yml
@@ -1,24 +1,0 @@
-language: php
-
-notifications:
-  email:
-    on_success: never
-    on_failure: change
-
-branches:
-  only:
-    - master
-
-php:
-  - 5.3
-  - 5.6
-
-env:
-  global:
-    - WP_CLI_BIN_DIR=/tmp/wp-cli-phar
-    - WP_CLI_CONFIG_PATH=/tmp/wp-cli-phar/config.yml
-
-before_script:
-  - bash bin/install-package-tests.sh
-
-script: ./vendor/bin/behat
