+++ behat.yml
@@ -1,0 +1,7 @@
+default:
+  suites:
+    default:
+      contexts:
+        - WP_CLI\Tests\Context\FeatureContext
+      paths:
+        - features
