--- features/extra/no-mail.php
@@ -1,6 +1,0 @@
-<?php
-
-function wp_mail() {
-	// do nothing
-}
-
