+++ wp-cli.yml
@@ -1,0 +1,2 @@
+require:
+  - command.php
